Begin refactoring event system and vulkan wrappers
Events are now handled by discrete observer classes (KeyHandler, MouseHandler) that can be utilized via composition, as opposed to the monolithic observer class that utilized inheritance in the previous system.

Each observer now subscribes to specific event types via the EventHandler's template-based Attach/Detach system, providing better type safety and flexibility.

The Key and Mouse observers now store and update a map containing the current state of Keys/Buttons of the window their EventHandler is attached to. This ensures that the map also registers keys that are in the GLFW_REPEAT state, which is not handled by glfwGetKey(). The Mouse observer also stores the previous mouse position for convenience and provides methods to set the cursor mode, as well as enable/disable raw mouse motion if supported by the system.

Slowly working on refactoring the vulkan/rendering systems. The main goal at the moment is abstracting all scene, object, mesh, camera, and similar classes from the vulkan wrappers so that they are all agnostic of the actual rendering implementations.

Usage of the previous event implementation has been stripped out, and testing is required for the new system. Testing will likely happen after the rendering refactor is closer to finalization, as that is the top priority.